增加渐变效果

+++ src/Color.h
@@ -3,13 +3,13 @@
 
 typedef struct
 {
-    int x;
-    int y;
+    double x;
+    double y;
 } Point;
 
 
-extern int get_color_line(Point _begin, Point _target, int _x);
+extern double get_color_line(Point _begin, Point _target, double _x);
 
-extern int get_color(int begin_color, int target_color, int mini, int maxi, int _x);
+extern int get_color(int begin_color, int target_color, double mini, double maxi, double _x);
 
 #endif
